+++ resources/services/lists/lists.go
@@ -1,11 +1,14 @@
 package lists
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"strings"
 
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/apache/arrow/go/v14/arrow"
+	"github.com/cloudquery/plugin-sdk/v4/schema"
+	"github.com/cloudquery/plugin-sdk/v4/types"
 	"github.com/koltyakov/cq-source-sharepoint/internal/util"
 	"github.com/koltyakov/gosip/api"
 	"github.com/rs/zerolog"
@@ -15,20 +18,12 @@ import (
 type Lists struct {
 	sp     *api.SP
 	logger zerolog.Logger
-
-	TablesMap map[string]Model // normalized table name to table metadata (map[CQ Table Name]Model)
-}
-
-type Model struct {
-	URI  string
-	Spec Spec
 }
 
 func NewLists(sp *api.SP, logger zerolog.Logger) *Lists {
 	return &Lists{
-		sp:        sp,
-		logger:    logger,
-		TablesMap: map[string]Model{},
+		sp:     sp,
+		logger: logger,
 	}
 }
 
@@ -65,10 +60,6 @@ func (l *Lists) GetDestTable(listURI string, spec Spec) (*schema.Table, error) {
 	}
 
 	fieldsData := fields.Data()
-	model := &Model{
-		URI:  listURI,
-		Spec: spec,
-	}
 
 	// ToDo: Rearchitect table construction logic
 	for _, prop := range spec.Select {
@@ -76,7 +67,7 @@ func (l *Lists) GetDestTable(listURI string, spec Spec) (*schema.Table, error) {
 		table.Columns = append(table.Columns, col)
 	}
 
-	l.TablesMap[table.Name] = *model
+	table.Resolver = l.Resolver(listURI, spec, table)
 
 	return table, nil
 }
@@ -101,19 +92,31 @@ func (l *Lists) getDestCol(prop string, tableName string, spec Spec, fieldsData 
 		fieldAlias = a
 	}
 
+	valueResolver := func(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+		value := util.GetRespValByProp(resource.Item.(map[string]any), prop)
+		if c.Type == arrow.BinaryTypes.String {
+			if value != nil {
+				value = fmt.Sprintf("%v", value)
+			}
+		}
+		return resource.Set(c.Name, value)
+	}
+
 	// Props is not presented in list's fields
 	if field == nil {
 		return schema.Column{
 			Name:        util.NormalizeEntityName(fieldAlias),
 			Description: prop,
 			Type:        typeFromPropName(prop),
+			Resolver:    valueResolver,
 		}
 	}
 
 	field.InternalName = fieldAlias
 	col := l.columnFromField(field, tableName)
-	col.CreationOptions.PrimaryKey = prop == "ID" // ToDo: Decide on ID cunstruction logic: use ID/UniqueID/Path+ID
+	col.PrimaryKey = prop == "ID" // ToDo: Decide on ID cunstruction logic: use ID/UniqueID/Path+ID
 	col.Description = prop
+	col.Resolver = valueResolver
 
 	return col
 }
@@ -151,32 +154,32 @@ func (l *Lists) columnFromField(field *api.FieldInfo, tableName string) schema.C
 
 	switch field.TypeAsString {
 	case "Text", "Note", "ContentTypeId":
-		c.Type = schema.TypeString
+		c.Type = arrow.BinaryTypes.String
 	case "Integer", "Counter":
-		c.Type = schema.TypeInt
+		c.Type = arrow.PrimitiveTypes.Int32
 	case "Currency":
-		c.Type = schema.TypeFloat
+		c.Type = arrow.PrimitiveTypes.Float32
 	case "Number":
-		c.Type = schema.TypeFloat
+		c.Type = arrow.PrimitiveTypes.Float32
 	case "DateTime":
-		c.Type = schema.TypeTimestamp
+		c.Type = arrow.FixedWidthTypes.Timestamp_us
 	case "Boolean", "Attachments":
-		c.Type = schema.TypeBool
+		c.Type = arrow.FixedWidthTypes.Boolean
 	case "Guid":
-		c.Type = schema.TypeUUID
+		c.Type = types.UUID
 	case "Lookup", "User":
-		c.Type = schema.TypeInt
+		c.Type = arrow.PrimitiveTypes.Int32
 	case "LookupMulti", "UserMulti":
-		c.Type = schema.TypeIntArray
+		c.Type = arrow.ListOf(arrow.PrimitiveTypes.Int32)
 	case "Choice":
-		c.Type = schema.TypeString
+		c.Type = arrow.BinaryTypes.String
 	case "MultiChoice":
-		c.Type = schema.TypeStringArray
+		c.Type = arrow.ListOf(arrow.BinaryTypes.String)
 	case "Computed":
-		c.Type = schema.TypeString
+		c.Type = arrow.BinaryTypes.String
 	default:
 		logger.Warn().Str("type", field.TypeAsString).Int("kind", field.FieldTypeKind).Str("field_title", field.Title).Str("field_id", field.ID).Msg("unknown type, assuming JSON")
-		c.Type = schema.TypeString
+		c.Type = arrow.BinaryTypes.String
 	}
 
 	c.Name = util.NormalizeEntityName(field.InternalName)
@@ -184,9 +187,9 @@ func (l *Lists) columnFromField(field *api.FieldInfo, tableName string) schema.C
 	return c
 }
 
-func typeFromPropName(prop string) schema.ValueType {
+func typeFromPropName(prop string) arrow.DataType {
 	if strings.HasSuffix(prop, "/Id") && prop != "ParentList/Id" {
-		return schema.TypeInt
+		return arrow.PrimitiveTypes.Int32
 	}
-	return schema.TypeString
+	return arrow.BinaryTypes.String
 }